Image_optimizer/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
install==1.3.5
Pillow==9.1.0
requests==2.27.1
urllib3==1.26.9
